append proxy owners

+++ server/controllers/proxy.js
@@ -1,10 +1,11 @@
+const { AdminRole } = require("../config/const");
 const ProxyService = require("../services/proxy");
-const { sendResult, sendError } = require("../utils/resp");
+const { sendResult, sendError, ApiError } = require("../utils/resp");
 
 const handleLoadProxies = async (req, res) => {
     try {
         const { page, pageSize } = req.query;
-        const [proxies, proxiesCount] = await ProxyService.loadProxies({ page, pageSize: pageSize || "10" })
+        const [proxies, proxiesCount] = await ProxyService.loadProxies(req.manager, { page, pageSize: pageSize || "10" })
         sendResult(res, { proxies, proxiesCount })
     } catch (error) {
         sendError(res, error)
@@ -14,7 +15,7 @@ const handleLoadProxies = async (req, res) => {
 const handleAddProxies = async (req, res) => {
     try {
         const { proxies, deadline } = req.body;
-        await ProxyService.addProxies(proxies, deadline)
+        await ProxyService.addProxies(req.manager, proxies, deadline)
         sendResult(res)
     } catch (error) {
         sendError(res, error);
@@ -23,9 +24,8 @@ const handleAddProxies = async (req, res) => {
 
 const handleClearProxies = async (req, res) => {
     try {
-        await ProxyService.clearProxies();
+        await ProxyService.clearProxies(req.manager);
         sendResult(res);
-
     } catch (error) {
         sendError(res, error);
     }
@@ -34,6 +34,11 @@ const handleClearProxies = async (req, res) => {
 const handleChangeProxyStatus = async (req, res) => {
     try {
         const { id, status } = req.body
+        const proxy = await ProxyService.findById(id)
+        if (!proxy)
+            throw new ApiError("The specified proxy doesn't exist.")
+        if (req.manager.role != AdminRole.MANAGER && req.manager._id.toString() != proxy.owner.toString())
+            throw new ApiError("The specified proxy can be changed only by owner.")
         await ProxyService.setProxyStatus(id, status)
         sendResult(res)
     } catch (error) {
@@ -45,6 +50,11 @@ const handleSetProxyStatus = async (req, res) => {
     try {
         const { id } = req.params
         const { status } = req.body
+        const proxy = await ProxyService.findById(id)
+        if (!proxy)
+            throw new ApiError("The specified proxy doesn't exist.")
+        if (req.manager.role != AdminRole.MANAGER && req.manager._id.toString() != proxy.owner.toString())
+            throw new ApiError("The specified proxy can be changed only by owner.")
         await ProxyService.setProxyStatus(id, status)
         sendResult(res)
     } catch (error) {
@@ -55,6 +65,11 @@ const handleSetProxyStatus = async (req, res) => {
 const handleDeleteProxy = async (req, res) => {
     try {
         const { id } = req.params
+        const proxy = await ProxyService.findById(id)
+        if (!proxy)
+            throw new ApiError("The specified proxy doesn't exist.")
+        if (req.manager.role != AdminRole.MANAGER && req.manager._id.toString() != proxy.owner.toString())
+            throw new ApiError("The specified proxy can be deleted only by owner.")
         await ProxyService.deleteProxy(id);
         sendResult(res);
     } catch (error) {
